added custom geometry
// const positionArray = new Float32Array(9)
// const positionArray = new Float32Array([......

The first is for if you want to come back to it later. You're saying its 9 vertices. 3 sets of coordinates vertexes. xyz.

The second is when you have all coordinates and want to pass them in at once. Formatted this way to highlight that there are three vertices with x,y,z coordinates. (Think of a triangle)