expanding parameter for spin
the prev gui mods were on the mesh and material which is why they are below the mesh and material. The color one is associated with the parameter const we made. We are expanding on that const to also assign a spin.

we added a console log to confirm that it is doing something upon press.