The loading manager
the loading manager is a threejs class so we can load it here

If we want to mutualize the ethe loading events. Something like the global loading process or to be informed once everything has loaded.

We use our loading manager in the paramters of our textureLoader to get a global output of info.