Using Math.PI
Math.PI on rotation is like this:

picture a matriox with an x and y with 4 quadrants.

imagine starting at x = 1, y = 0
imagine moving in a counter clockwise motion.

if you want to rotate a half circle, its Math.PI

If I want to do a full circle is Math.PI * 2

If I want to do a 8th of a rotation: Math.PI / 4