mipmapping repeat textures
If you look at the cube's top face while this face is almost hidden, you'll see a very blurry texture.

Mipmapping (or "mip mapping" with a space) is a technique that consists of creating half a smaller version of a texture again and again until you get a 1x1 texture. All those texture variations are sent to the GPU, and the GPU will choose the most appropriate version of the texture.

Three.js and the GPU already handle all of this, and you can just set what filter algorithm to use.

meaning 1/2 then 1/4 then 1/8, etc. depending on how many of the pixels we can see. The GPU is doing this.

This is good.

There are two types of filter algorithms: the minification filter and the magnification filter.