min and mag filter to fix gradient
If you test this, you'll see that the cartoon effect doesn't work anymore. That is because the gradient texture we used is tiny, and the pixels of that texture are blended. Yes, it's a matter of minFilter, magFilter, and mipmapping like we saw in the Textures lesson.

To fix this, we can simply change the minFilter and magFilter to THREE.NearestFilter.

Using THREE.NearestFilter means that we are not using the mip mapping, we can deactivate it with gradientTexture.generateMipmaps = false: