create debug for tweaks
We created a debug UI managing rougness and metalness

First, we must add the Dat.GUI dependency to our project. In the terminal, on the project folder (where the server should be currently running), use the following command:

Then, on top of your code, import dat.gui (don't forget to re-launch the server with npm run dev if you stopped it):

You can now create an instance of it:

And add the tweaks (after creating the material):

And that's it. You can now change the metalness and the roughness as you like.