for Ambient Oclusion
We can simply add new attributes like we did on the Geometries lesson and use the default uv attribute. In more simple terms, we duplicated the uv attribute.

Call this new attribute uv2:

You can now add the aoMap using the doorAmbientOcclusionTexture texture and control the intensity using the aoMapIntensity property:

Bonus: added it to the Gui