Environment map
The environment map is like an image of what's surrounding the scene. You can use it to add reflection or refraction to your objects. It can also be used as lighting information.

We haven't covered it yet, but you can use it with many of the materials we saw.

First, let's setup a very simple MeshStandardMaterial with the debug UI as we did earlier: