update shadowmap render size
shadowmaps have their own width and height. To improve the shadow quality, we have to change the default height and width. Keep in mind this has to be a factor of 2 for mipmapping.
(using console.log we can investigate the directionalLight's shadow properties and see the default mapsize is 512x512)
mapSize: Vector2 {x: 512, y: 512, isVector2: true}