Baking shadows
Three.js shadows can be very useful if the scene is simple, but it might otherwise become messy.

A good alternative is baked shadows. We talk about baked lights in the previous lesson and it is exactly the same thing. Shadows are integrated into textures that we apply on materials.

add texture loader

import texture "baked shadow" from the static folder

And finally, instead of using a MeshStandardMaterial on the plane, we'll use a simple MeshBasicMaterial with the bakedShadow as map: