Baking shadows alternative
The texture is a simple halo. The white part will be visible and the black part will be invisible.

Then, we move that shadow with the sphere. it will be a plane right below the sphere and right above the floor plane

First, let's remove the previous baked shadow by putting back the MeshStandardMaterial on the plane:

Then, we can load a basic shadow texture located in /static/textures/backedShadow.jpg.

We can create the shadow by using a simple plane that we rotate and place slightly above the floor. The material must be black but with the shadow texture as the alphaMap. Don't forget to change transparent to true, and to add the mesh to the scene:

There you go, a not so realistic but very performant shadow.