Door baseline
We will use a simple plane for the door as we are going to use the beautiful door texture we used in a previous lesson.

We don't know yet if the plane has the right size, but we can fix that later when we have the textures working.

As you can see, we move the door on the z axis to stick it to the wall but we also added 0.01 units. If you don't add this small value, you'll have a bug we already saw in a previous lesson called z-fighting. Z-fighting happens when you have two faces in the same position (or very close). The GPU doesn't know which one is closer than the other, and you get some strange visual pixel fighting.