Add graves
The idea is to place the graves randomly on a circle around the house.

The idea is to place the graves randomly on a circle around the house.

Like in the 3D Text lesson where we created multiple donuts with one geometry and one material, we are going to create one BoxGeometry and one MeshStandardMaterial that will be shared amongst every graves:

Finally, let's loop and do some mathematics to position a bunch of graves around the house.

We are going to create a random angle on a circle. Remember that a full revolution is 2 times π. Then we are going to use that angle on both a sin(...) and a cos(...). This is how you position things on a circle when you have the angle. And finally we also multiply those sin(...) and cos(...) results by a random value because we don't want the graves to be positioned on a perfect circle.