Fog
In horror movies, they always use fog. The good news is that Three.js supports it already with the Fog class.

The first parameter is the color, the second parameter is the near (how far from the camera does the fog start), and the third parameter is the far (how far from the camera will the fog be fully opaque).

To activate the fog, add the fog property to the scene: