implement door textures
Then we can apply all those textures to the door material. Don't forget to add more subdivisions to the PlaneGeometry, so the displacementMap has some vertices to move. Also, add the uv2 attribute to the geometry for the aoMap as we did in the Materials lesson.

You can access the door's geometry by using mesh.geometry: