add walls using textures
Let's do the same for the walls using the textures on the /static/textures/bricks/ folder. We don't have as many textures as for the door, but it's not a problem. We don't need an alpha texture, and the wall has no metal in it, so we don't need a metalness texture either.

Load the textures: Then we can update our MeshStandardMaterial for the wall. Don't forget to remove the color and add the uv2 attribute for the ambient occlusion.