optimize shadows
The scene looks much better with these shadows, but we should optimize them anyway.

A good thing would be to go through each light, create camera helpers on the light.shadowMap.camera, and make sure the near, the far, the amplitude or the fov fit nicely. But instead, let's use the following values that should be just right.

We can also reduce the shadow map render sizes to improve performances: