Changing the particles with Custom geometry
To create a custom geometry, we can start from a BufferGeometry, and add a position attribute as we did in the Geometries lesson. Replace the SphereGeometry with custom geometry and add the 'position' attribute as we did before.

But creating new geometries can be really hard. There is a lot to remember, like declaring every vertex, and communicating those points in an array of x1,y1,z1,x2,y2,z2,etc.