center stars (instead of starting at 0)
We want to move positions of [i] to the left by half. so we take Math.random() and subtract .5

math.random generates random integers between 0 and 1. So about half should be under .5, so subtracting .5 will result in about half being a negative number.

Multiplying these by 10 spreads out the stars by a magnitude of 1.