Implement random colors using vertex colors process
We can have a different color for each particle. We first need to add a new attribute named color as we did for the position. A color is composed of red, green, and blue (3 values), so the code will be very similar to the position attribute. We can actually use the same loop for these two attributes:

Be careful with singulars and plurals.

To activate those vertex colors, simply change the vertexColors property to true:

note: The main color of the material still affects these vertex colors. Feel free to change that color or even comment it.