Get the angle and position on branches
We can use Math.cos(...) and Math.sin(...) to position the particles on those branches. We first calculate an angle with the modulo (%), divide the result by the branches count parameter to get an angle between 0 and 1, and multiply this value by Math.PI * 2 to get an angle between 0 and a full circle. We then use that angle with Math.cos(...) and Math.sin(...) for the x and the z axis and we finally multiply by the radius: