Randomness
Those particles are perfectly aligned. We need randomness. But what we truly need is spread stars on the outside and more condensed star on the inside.

Create a randomness parameter:

Now create a random value for each axis with Math.random(), multiply it by the radius and then add those values to the postions: