Randomness
Those particles are perfectly aligned. We need randomness. But what we truly need is spread stars on the outside and more condensed star on the inside.

Create a randomness parameter:

Now create a random value for each axis with Math.random(), multiply it by the radius and then add those values to the postions:

It's working but it's not very convincing, right? And we can still see the pattern. To fix that, we can use Math.pow() to crush the value. The more power you apply, the closest to 0 it will get. The problem is that you can't use a negative value with Math.pow(). What we will do is calculate the power then multiply it by -1 randomly.

First create the power parameter:

Then apply the power with Math.pow() and multiply it by -1 randomly: