Normalize (shown in console log)
This example of a normalized vector isn't very relevant because we could have set 1 instead of 10, but if we change the values, we will still have the normalize() method making sure that the vector is 1 unit long.

Used console.log(rayDirection.length())
 before and after to display the length. outputs were 10 and then 1 due to normalization. Helps make exact positioning easier to manage.

Here, the ray position supposedly start a little on the left in our scene, and the direction seems to go to the right. Our ray should go through all the spheres.