turn blue on intersect
JavaScript
Yet again, we don't really need to normalize the rayDirection because its length is already 1. But it's good practice to keep the normalize() in case we change the direction.

We also put the array of objects to test in a variable objectsToTest. That will get handy later.

If you look at the console, you should get an array with intersections in it, and those intersections keep changing depending on the spheres' positions.

We can now update the material of the object property for each item of the intersects array: