turn red if no intersect, turn blue during intersect, and go back to red
Unluckily, they all go blue but never go back red. There are many ways to turn the objects that didn't intersect back to red. What we can do is turn all the spheres red and then turn the ones that intersect blue:

We are looping on a for to set object.material.color to blurple but if it's intersecting (after!) then it will turn blue. If its not intersecting it will keep turning blurple.