use raycaster with mouse / hovering I
establish coordinates between -1 -> 1 for both the x and y.

As we said earlier, we can also use the raycaster to test if an object is behind the mouse. In other words, if you are hovering an object.

Mathematically speaking, it's a little complex because we need to cast a ray from the camera in the mouse's direction, but, fortunately, Three.js is doing all the heavy lifting.

For now, let's comment the code related to the raycaster in the tick function.

First, let's handle hovering.

To begin with, we need the coordinates of the mouse. We cannot use the basic native JavaScript coordinates, which are in pixels. We need a value that goes from -1 to +1 in both the horizontal and the vertical axis, with the vertical coordinate being positive when moving the mouse upward.

This is how WebGL works and it's related to things like clip space but we don't need to understand those complex concepts.

Examples:

The mouse is on the top left of the page: -1 / 1
The mouse is on the bottom left of the page: -1 / - 1
The mouse is on the middle vertically and at right horizontally: 1 / 0
The mouse is in the center of the page: 0 / 0
First, let's create a mouse variable with a Vector2, and update that variable when the mouse is moving: