use raycaster with mouse / hovering II
Look at the logs and make sure that the values match the previous examples.

We could cast the ray in the mousemove event callback, but it's not recommend because the mousemove event might be triggered more than the frame rate for some browsers. Instead, we will cast the ray in the tick function as we did before.

To orient the ray in the right direction, we can use the setFromCamera() method on the Raycaster. The rest of the code is the same as before. We just update the objects materials to red or blue if they intersect or not: (The spheres should become blue if the cursor is above them.

)