Mouse enter and mouse leave events
Mouse events like 'mouseenter', 'mouseleave', etc. aren't supported either. If you want to be inform when the mouse "enters" an object or "leaves" that object, you'll have to do it by yourself.

What we can do to reproduce the mouseenter and mouseleave events, is to have a variable containing the currently hovered object.

If there is one object intersecting, but there wasn't one before, it means a mouseenter has happened on that object.

If no object intersects, but there was one before, it means a mouseleave has happened.

We just need to save the currently intersecting object:
let currentIntersect = null

And then, test and update the currentIntersect variable: