notes
First, we need to create a Cannon.js World:

We could make a WebGL experience in space where there is no gravity, but let's keep our feet on Earth and add gravity. You can change this value with the gravity property, which is a Cannon.js Vec3.

We used - 9.82 as the value because it's the gravity constant on earth, but you can use any other value if you want things to fall slower or if your scene happens on Mars.

Cannon.js Vec3 is just like Three.js Vector3. It has x, y, and z properties, but also a set(...) method:

Because we already have a sphere in our scene, let's create a sphere inside our Cannon.js World.

To do that, we must create a Body. Bodies are simply objects that will fall and collide with other bodies.

Before we can create a Body, we must decide on a shape. There are many available primitive shapes like Box, Cylinder, Plane, etc. We will go for a Sphere with the same radius as our Three.js sphere:

Then we can create our Body and specify a mass and a position:

Finally, we can add the Body to the world with addBody(...):

Nothing is happening because we still need to update our Cannon.js world and update our Three.js sphere accordingly.