Update the Cannon.js world and the Three.js scene
To update our World, we must use the step(...). The code behind this method is hard to understand, and we won't explain it in this lesson, but you can find more about it in this article.

For it to work, you must provide a fixed time step, how much time passed since the last step, and how much iterations the world can apply to catch up with a potential delay.

We won't explain what is a time step, but because we want our experience to run at 60fps, we are going to use 1 / 60. Don't worry, the experience will work at the same speed on devices with higher and lower frame rates.

It's up to you for the number of iterations, but it's not that important if the experience is running smoothly. Let's use 3.

For the delta time, it's a little harder. We need to calculate how much time has passed since the last frame. Do not use the getDelta() method from the Clock class. You won't get the intended result, and you'll mess up with the class's internal logic.

To get the right delta time, we need to subtract the elapsedTime from the previous frame to the current elapsedTime: