Contact Material
As you can see, the ball doesn't bounce much. That is the default behavior, and we can change that with Material (Not the Material from Three.js) and ContactMaterial.

A Material is just a reference. You can give it a name and associate it with a Body. The idea is to create a Material for each type of material you have in your scene.

Suppose everything in your world is made out of plastic. In that case, you only have to create one material and name it 'default' or 'plastic'. If you have multiple types of materials in your scene, let's say one material for the floor and one for the ball. Then, you should create various Material and give them names like 'concrete' and 'plastic'.

You could have called them 'ground' and 'ball'. Still, if you want to use the same materials for walls and other objects like cubes, it could be inconvenient to use a material named 'ground'.

Before you create the sphere and the floor, create these two Material: