contact materials - global default material
We get the same result.

We can go even further by setting our material as the default one of our World. To do that, simply assign the defaultContactMaterial to the world's defaultContactMaterial property:

world.defaultContactMaterial = defaultContactMaterial

We can now remove or comment the material assignation of the sphereBody and floorBody.