Apply forces
There are many ways to apply forces to a Body:

applyForce to apply a force to the Body from a specified point in space (not necessarily on the Body's surface) like the wind that pushes everything a little all the time, a small but sudden push on a domino or a greater sudden force to make an angry bird jump toward the enemy castle.
applyImpulse is like applyForce but instead of adding to the force that will result in velocity changes, it applies directly to the velocity.
applyLocalForce is the same as applyForce but the coordinates are local to the Body (meaning that 0, 0, 0 would be the center of the Body).
applyLocalImpulse is the same as applyImpulse but the coordinates are local to the Body.
Because using "force" methods will result in velocity changes, let's not use "impulse" methods

Let's use applyLocalForce(...) to apply a small impulse on our sphereBody at the start:

sphereBody.applyLocalForce(new CANNON.Vec3(150, 0, 0), new CANNON.Vec3(0, 0, 0))

You can see the ball bounce to the right and roll.

Now let's use the applyForce(...) to apply some wind. Because the wind is permanent, we should apply this force to each frame before updating the World. To correctly apply this force, the point should be the sphereBody.position: