Automate with functions part 1
we are going to create a function to automate the creation of spheres.

To start with, let's improve how we create spheres with a function that will add both the Three.js and the Cannon.js versions.

As parameters of this function, we will only pass the radius and the position, but feel free to add other parameters such as mass, material, subdivisions, etc.
Now we can create the Three.js Mesh:

And the Cannon.js Body:

We can remove the previously created sphere and call the createSphere(...) (after creating the Cannon.js world and the Three.js scene). Don't forget to remove the sphere update in the tick() function:
As you can see, the position doesn't have to be a Three.js Vector3 or a Cannon.js Vec3 and we can simply use an object with x, y and z properties (luckily for us).

You should see the sphere floating above the floor, but unfortunately, it's not moving anymore. And this is perfectly normal because we removed the code that was taking the Cannon.js Body position to apply it to the Three.js Mesh position.