raining random spheres
And then add functions to it when needed (after creating the createSphere function):

Finally, we can add this new createSphere property to Dat.GUI:

If you click on the newly created createSphere button, you should see spheres staking one on each other. That is due to the sphere popping on the exact same position. Let's add some randomness:

It's raining spheres!

Try not to burn your computer; this code needs optimization.