Proper Boxes
Now that our spheres are working well, let's do the same process but for the boxes.

To create a box, we must use a BoxGeometry and a Box shape. Be careful; the parameters aren't the same. A BoxGeometry needs a width, a height, and a depth. In the meantime, a Box shape needs a halfExtents. It is represented by a Vec3 corresponding to a segment that starts at the center of the box and joining one of that box corners:

Don't forget to remove the first createSphere(...) call, or you'll have both the sphere and the box created in the same position simultaneously, which might get messy.