Brad Phase
as it stands, every object is testing collision with every other object. This is bad for performance. This is called the broad phase. Currently CANNON is using naive broad phase, but there are other solutions like grid broad phase or sweep and prune SAPBroadphase (which will throw errors if objects move too fast)